uppsrc: Fixed many Win64 warnings
git-svn-id: svn://ultimatepp.org/upp/trunk@5821 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/SSL/Util.cpp
@@ -6,7 +6,7 @@ String SslBuffer::Get() const
 {
 	if(IsEmpty())
 		return String::GetVoid();
-	return String(buf_mem->data, buf_mem->length);
+	return String(buf_mem->data, (int)buf_mem->length);
 }
 
 bool SslBuffer::Grow(int length)
@@ -50,7 +50,7 @@ String SslStream::GetResult() const
 	BIO_get_mem_ptr(bio, &bm);
 	if(!bm)
 		return String::GetVoid();
-	return String(bm->data, bm->length);
+	return String(bm->data, (int)bm->length);
 }
 
 bool SslKey::Load(const String& data)
